utils: Add bench_stats()
Adds bench_stats(), a function to summarise results of a run.
Upcoming test update will perform tests in loops. This function provides
a nice way to get maximum, minimum and average from an array of
uint32_t.

An uint64_t is used to sum all entries of input array - that should be
enough for fairly large arrays. This assumption may be revisited in the
future.

Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>