semaphore: Reduce memory requirements
Instead of using ITERATION sized arrays to track the times for each
iteration of the test, a smaller structure is used to track the
average, minimum, maximum and total times. By reducing the memory
footprint, this allows for additional systems with less available
memory to utilize this benchmark utilility.

bench_utils update to provide helpers for using this new struct:
	bench_stats_reset: Reset struct to initial status
	bench_stats_update: Add new entry and update min, max and avg
			    values

Calibration of how long does it take to make time measurements is also
done during reset.

Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>