tests: Clean up "static" usage
For the upcoming feature that allows all tests to be run on a single
execution, the test entry point should not be static - as it will be
called from somewhere else.

Also, some non-static global variables for the tests were made static,
so they don't conflict with each other when all tests are linked
together.

Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>