tests: Add possibility to run all tests at once
If cmake option "-DTEST=<test>" is not provided, all tests are built and
run.

The main() function for each test is now enclosed on a ifdef block that
controls if it should be built (when only running a single test) or not.

bench_all.c file has the main() function for the collective test run.

Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>