freertos: Implement mutex and thread related functions
Two notes: first is that FreeRTOS priority increase with number
increased, opposed to Zephyr. A simple conversion is made to
keep using Zephyr's order on tests.

Second is that there's no concept of creating a thread and preventing it
from being scheduled, to start it later. So only `bench_thread_create`
is implemented.

Signed-off-by: Ederson de Souza <ederson.desouza@intel.com>