build: Re-organize directory structure - part 4
Updates the Zephyr build scripts so that parts that are zephyr
specific are contained under src/zephyr.

At this point, re-organizing the directory structure for
Zephyr is complete, and it is to be the exemplar for the
addition of other RTOSes.

src/common
src/${RTOS}
src/${RTOS}/arch
src/${RTOS}/timer
src/${RTOS}/CMakeLists.txt
src/${RTOS}/${RTOS}.cmake

Signed-off-by: Peter Mitsis <peter.mitsis@intel.com>