reporting: Use "n/a" if test is not valid for RTOS
Some RTOSes distinguish between creating, starting and spawning
a thread. This has a direct impact on whether benchmark should
be reported for creating, starting and spawning threads. To
assist with this, two new macros have been created:

    RTOS_HAS_THREAD_SPAWN
    RTOS_HAS_THREAD_CREATE_START

If defined to a non-zero value, then the RTOS in question supports
that type of thread creation.

Signed-off-by: Peter Mitsis <peter.mitsis@intel.com>