Remove calibration adjustment
It has been observed on QEMU that the timings can be somewhat
inaccurate (varying loads on the host). This in turn can lead
to anomalous reported values. As the time to obtain a timestamp
is generally much much less than what is being measured, the
calibration (estimated time spent getting a timestamp) is being
removed in favor of more consistent reporting.

Signed-off-by: Peter Mitsis <peter.mitsis@intel.com>